modified homework_3 and homework_4
