+++ src/homework_3/Hovhannes_Gspeyan/Task_2.java
@@ -7,12 +7,12 @@ public class Task_2 {
     public static void main(String[] args) {
 //        printNumbers(10);            // Question 1 -> try with different integer values instead of 10 only
 //        sumNumbers();                // Question 2
-        table();                     // Question 3
+//        table();                     // Question 3
 //        factorialDemo1();            // Question 4
 //        powerDemo();                 // Question 5
 //        reverseNumber();             // Question 6
 //        readSetIntegers();           // Question 7
-        testPrime();                 // Question 8
+//        testPrime();                 // Question 8
     }
