+++ src/homework_4/Hovhannes_Gspeyan/stack/ArrayStack.java
@@ -2,56 +2,41 @@ package homework_4.Hovhannes_Gspeyan.stack;
 
 public class ArrayStack {
 
-    int[] arr;
-    int size;
+    private int[] arr;
+    private int size;
 
-    ArrayStack(int capacity) {
+    public ArrayStack(int capacity) {
         this.arr = new int[capacity];
     }
-    int size() {
-        return this.size;
-    }
-    boolean isEmpty() {
+    public boolean isEmpty() {
         return this.size() == 0;
-
     }
-    boolean isFully() {
+    public boolean isFully() {
         return this.size == arr.length;
     }
-    void push(int value) {
+    public void push(int value) {
         if (this.isFully()) {
             System.out.println("Error stack is fully! ");
             return;
         }
-        this.checkAndGrow(1);
+        this.checkAndGrow();
         this.arr[this.size++] = value;
     }
-    int pop() {
+    public int pop() {
         if (this.isEmpty()) {
             System.err.println("Error Stack is empty");
             return -1;
         }
         return this.arr[--this.size];
     }
-    void checkAndGrow(int count) {
-        if (this.size + count >= (arr.length * 75) / 100) {
+    private int size() {
+        return this.size;
+    }
+    private void checkAndGrow() {
+        if (this.size + 1 >= (arr.length * 75) / 100) {
             int[] newArr = new int[arr.length * 2];
-            for (int i = 0; i < arr.length; i++) {
-                newArr[i] = arr[i];
-            }
+            System.arraycopy(arr, 0, newArr, 0, arr.length);
             arr = newArr;
         }
     }
-    public static void main(String[] args) {
-        ArrayStack stack = new ArrayStack(10);
-        int value = 0;
-        for (int i = 0; i < 40; i++) {
-            stack.push(value++);
-        }
-        while (!stack.isEmpty()) {
-            System.out.println(stack.pop());
-        }
-    }
-
-
 }
